+++ fractals/src/main/java/me/catmousedog/fractals/fractals/types/normalized/NormalizedJulia.java
@@ -111,7 +111,7 @@ public final class NormalizedJulia extends Fractal {
 					jy = t[1];
 					update();
 					if (settings.isRender_on_changes())
-						canvas.render();
+						jpi.renderNow();
 				}
 			}
 		};